Headers/Tags update, FAQ subpage formatting fixed, subpage tag removed, FAQ and Troubleshooting combined
Been through a lot of iterations of how this should look and I'm the happiest with this so far. Let's get something even better next week. Main point of contrition was combining the FAQ and Troubleshooting sections but there's not enough pages when sorted for them to really be separate. More Troubleshooting pages coming soon but the line between "this is so specific it belongs on the page its on" as opposed to "this is helpful enough to be a wider page" is surprisingly difficult to find.