FAQ, Visual changes, Sidebar re-org, Troubleshooting, Deprecated deletions, Hero page (#280)
* faq-troubleshooting branch open, faq pages added

* relinked troubleshooting pages

* Catching up with local. Sidebar changes, irrelevant pages deleted

* many visual changes, hero page added, sidebar further reorganized

* fixed lint

* Untracked template file

* removing hero splash for a future update

* removed doxybook

* removed template.md

* Headers/Tags update, FAQ subpage formatting fixed, subpage tag removed, FAQ and Troubleshooting combined

Been through a lot of iterations of how this should look and I'm the happiest with this so far. Let's get something even better next week. Main point of contrition was combining the FAQ and Troubleshooting sections but there's not enough pages when sorted for them to really be separate. More Troubleshooting pages coming soon but the line between "this is so specific it belongs on the page its on" as opposed to "this is helpful enough to be a wider page" is surprisingly difficult to find.

* Moved API pages up one level to docs. Fixed broken anchors.

* removed /api/ reference from footer

* fixed api/central references